booking/requirements.txt
ariadne==0.26.2
Flask==3.1.2
requests==2.32.5
grpcio==1.76.0
grpcio-tools==1.76.0
protobuf==6.32.1
graphql-core==3.2.5
pymongo==4.10.1
